+++ tensorrt_llm/_torch/custom_ops/__init__.py
@@ -1,3 +1,4 @@
+from ..cute_dsl_utils import IS_CUTLASS_DSL_AVAILABLE
 from ..flashinfer_utils import IS_FLASHINFER_AVAILABLE
 from ..modules.attention import attn_custom_op_inplace, mla_custom_op_inplace
 from .cpp_custom_ops import _register_fake
@@ -15,6 +16,7 @@ __all__ = [
     'matmul_to_ub',
     'attn_custom_op_inplace',
     'mla_custom_op_inplace',
+    'IS_CUTLASS_DSL_AVAILABLE',
 ]
 
 if IS_FLASHINFER_AVAILABLE:
@@ -27,4 +29,10 @@ if IS_FLASHINFER_AVAILABLE:
         'flashinfer_rmsnorm',
         'flashinfer_fused_add_rmsnorm',
         'flashinfer_apply_rope_with_cos_sin_cache_inplace',
+    ]
+
+if IS_CUTLASS_DSL_AVAILABLE:
+    from .cute_dsl_custom_ops import cute_dsl_nvfp4_gemm_blackwell
+    __all__ += [
+        'cute_dsl_nvfp4_gemm_blackwell',
     ]
