+++ tensorrt_llm/_torch/cute_dsl_utils.py
@@ -1,0 +1,16 @@
+import platform
+import traceback
+
+from ..logger import logger
+
+IS_CUTLASS_DSL_AVAILABLE = False
+
+if platform.system() != "Windows":
+    try:
+        logger.info(f"cutlass dsl is available")
+        IS_CUTLASS_DSL_AVAILABLE = True
+    except ImportError:
+        traceback.print_exc()
+        print(
+            "cutlass dsl is not installed properly, please try pip install nvidia-cutlass-dsl"
+        )
