+++ tests/unittest/_torch/thop/parallel/test_fp4_linear.py
@@ -6,6 +6,7 @@ from utils.util import skip_pre_blackwell
 
 import tensorrt_llm.quantization.utils.fp4_utils as fp4_utils
 from tensorrt_llm._torch.autotuner import autotune
+from tensorrt_llm._torch.custom_ops import IS_CUTLASS_DSL_AVAILABLE
 from tensorrt_llm._torch.modules.linear import Linear
 from tensorrt_llm._utils import get_sm_version
 from tensorrt_llm.models.modeling_utils import QuantAlgo, QuantConfig
@@ -91,6 +92,8 @@ def pad_up(x, pad_size):
     get_sm_version() != 100,
     reason="This test is only supported in Blackwell architecture",
 )
+@pytest.mark.skipif(not IS_CUTLASS_DSL_AVAILABLE,
+                    reason="cutlass-dsl is not available")
 @pytest.mark.parametrize("dtype", [torch.bfloat16])
 @pytest.mark.parametrize("mnk", [(128, 7168, 16384), (128, 24576, 1536),
                                  (128, 2112, 7168), (128, 4096, 7168),
